+++ docs/changelog.rst
@@ -28,6 +28,8 @@ Features:
 Bugfix:
 
 * Fix incorrect arguments for the TagField's ``RelatedManager.set``
+* Upgrade select2 to fix composed characters (#138)
+* Fix select2 input where quotes in quoted tags could be escaped
 
 
 Thanks to: